# backend/resources/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MOVIC API",
    "version": "1.0.0",
    "description": "Documentação interativa da API do MOVIC."
  },
  "servers": [
    { "url": "http://localhost:8000/api" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ExerciseCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "category": { "type": "string" },
          "video_url": { "type": "string" }
        },
        "required": ["name"]
      },
...
    }
  },
  "paths": {
...
    "/teacher/exercises": {
      "get": {
        "summary": "Listar exercicios",
        "tags": ["Teacher"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Criar exercicio",
        "tags": ["Teacher"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExerciseCreate" } } }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/teacher/exercises/{exerciseId}": {
      "put": {
        "summary": "Atualizar exercicio",
        "tags": ["Teacher"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "exerciseId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExerciseCreate" } } }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Remover exercicio",
        "tags": ["Teacher"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "exerciseId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}